requirements.txt
pyspark==3.5.1
psycopg2-binary==2.9.9
pyyaml==6.0.2
apache-airflow==2.9.3
apache-airflow-providers-apache-spark==5.3.2
apache-airflow-providers-postgres==5.6.0
